--- core/agent_harness/integrations/__init__.py
@@ -1,15 +1,0 @@
-"""Agent-harness integration resolution helpers."""
-
-from core.agent_harness.integrations.resolution import (
-    IntegrationResolutionRequest,
-    IntegrationResolutionResult,
-    resolve_integrations,
-    resolve_integrations_with_metadata,
-)
-
-__all__ = [
-    "IntegrationResolutionRequest",
-    "IntegrationResolutionResult",
-    "resolve_integrations",
-    "resolve_integrations_with_metadata",
-]
