+++ tests/tools/investigation/stages/resolve_integrations/test_resolve_integrations_node.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from core.agent_harness.integrations.resolution import IntegrationResolutionResult
+from core.agent_harness.session.integration_resolution import IntegrationResolutionResult
 from tools.investigation.stages.resolve_integrations import node
